Remove using useAuthStore for protected route, moved transaction-api inside services/routes
